Backend/usabilidad_backend/requirements.txt
asgiref==3.3.1
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
Django==3.0.5
django-cors-headers==3.7.0
djangorestframework==3.12.2
dnspython==2.1.0
httplib2==0.19.1
idna==3.2
logzero==1.7.0
oauthlib==3.1.1
packaging==21.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.9.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
sqlparse==0.2.4
uritemplate==3.0.1
urllib3==1.26.6